Basic BT and start of Cover System
A simple behaviour tree allowing the enemy to detect, shoot, and chase the player. Work has been started on the cover system to make the AI's behaviour more complex and realistic; so far static waypoints have been added to the map, and the AI can detect the closest ones. Currently, work is being done to add score system based on the distance of each waypoint to the enemy.