Using A* to calculate distance to nearby waypoints
A* can now be used to calculate the distance to each nearby waypoint from the NPC.  An interval was introduced to the FindCoverSpot.cs in order to lower the computational load. Later on the cover will be calculated only when needed, not every frame or so.  Next, calculating scores will be added, possibly with a floating text displaying the cost above the waypoint.