Chasing and Shooting
The FSM has been expanded to be capable of making the enemy chase the player. An instance of KnownEnemiesBlackboard class is shared between the enemies on the same team to share information about the known enemies and their location; this way, even if the enemy doesn't see an enemy right now, it is still possible to chase one seen by its companions.

TODO: improve chasing state to use previousPosition to calculate where to look for the enemy, as well as add an option for the enemy to look around after reaching the target position.